requirements.txt
colorama==0.4.6
fastapi[all]==0.111.0
python-dotenv==1.0.1
pydantic-settings==2.3.1
sqlmodel==0.0.18
passlib[argon2]==1.7.4
itsdangerous==2.2.0
PyJWT==2.8.0
tzdata==2024.1
nanoid==2.0.0
requests==2.32.3
motor==3.5.1
